server/requirements.txt
# Adaptive RAG Server - Production Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0

# ML & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# OCR & PDF (for indexing pipeline only)
# Uncomment if running indexing:
# opencv-python>=4.8.0
# pdf2image>=1.16.0
# pillow>=10.0.0
# pytesseract>=0.3.10
